@@ -1,55 +1,65 @@
 {
   "name": "chatgpt-next-web",
-  "version": "1.9.3",
   "private": false,
-  "license": "Anti 996",
+  "license": "mit",
   "scripts": {
-    "dev": "yarn fetch && next dev",
-    "build": "yarn fetch && next build",
+    "dev": "next dev",
+    "build": "cross-env BUILD_MODE=standalone next build",
     "start": "next start",
     "lint": "next lint",
-    "fetch": "node ./scripts/fetch-prompts.mjs",
+    "export": "cross-env BUILD_MODE=export BUILD_APP=1 next build",
+    "export:dev": "cross-env BUILD_MODE=export BUILD_APP=1 next dev",
+    "app:dev": "yarn tauri dev",
+    "app:build": "yarn tauri build",
+    "prompts": "node ./scripts/fetch-prompts.mjs",
     "prepare": "husky install",
     "proxy-dev": "sh ./scripts/init-proxy.sh && proxychains -f ./scripts/proxychains.conf yarn dev"
   },
   "dependencies": {
-    "@hello-pangea/dnd": "^16.2.0",
-    "@microsoft/fetch-event-source": "^2.0.1",
+    "@fortaine/fetch-event-source": "^3.0.6",
+    "@hello-pangea/dnd": "^16.3.0",
     "@svgr/webpack": "^6.5.1",
     "@vercel/analytics": "^0.1.11",
     "emoji-picker-react": "^4.4.7",
     "fuse.js": "^6.6.2",
-    "mermaid": "^10.1.0",
-    "next": "^13.4.2",
+    "html-to-image": "^1.11.11",
+    "mermaid": "^10.2.3",
+    "nanoid": "^4.0.2",
+    "next": "^13.4.9",
     "node-fetch": "^3.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-markdown": "^8.0.5",
-    "react-router-dom": "^6.10.0",
+    "react-markdown": "^8.0.7",
+    "react-router-dom": "^6.14.1",
     "rehype-highlight": "^6.0.0",
-    "rehype-katex": "^6.0.2",
+    "rehype-katex": "^6.0.3",
     "remark-breaks": "^3.0.2",
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1",
     "sass": "^1.59.2",
     "spark-md5": "^3.0.2",
-    "use-debounce": "^9.0.3",
-    "zustand": "^4.3.6"
+    "use-debounce": "^9.0.4",
+    "zustand": "^4.3.8"
   },
   "devDependencies": {
-    "@types/node": "^18.14.6",
-    "@types/react": "^18.0.28",
+    "@tauri-apps/cli": "^1.4.0",
+    "@types/node": "^20.3.3",
+    "@types/react": "^18.2.14",
     "@types/react-dom": "^18.0.11",
     "@types/react-katex": "^3.0.0",
     "@types/spark-md5": "^3.0.2",
     "cross-env": "^7.0.3",
-    "eslint": "^8.36.0",
+    "eslint": "^8.44.0",
     "eslint-config-next": "13.2.3",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.0",
-    "lint-staged": "^13.2.0",
-    "prettier": "^2.8.7",
-    "typescript": "4.9.5"
+    "lint-staged": "^13.2.2",
+    "prettier": "^2.8.8",
+    "typescript": "4.9.5",
+    "webpack": "^5.88.1"
+  },
+  "resolutions": {
+    "lint-staged/yaml": "^2.2.2"
   }
 }
