@@ -1,15 +1,4 @@
-const COMMIT_ID: string = (() => {
-  try {
-    const childProcess = require("child_process");
-    return childProcess
-      .execSync('git log -1 --format="%at000" --date=unix')
-      .toString()
-      .trim();
-  } catch (e) {
-    console.error("[Build Config] No git or not from git repo.");
-    return "unknown";
-  }
-})();
+import tauriConfig from "../../src-tauri/tauri.conf.json";
 
 export const getBuildConfig = () => {
   if (typeof process === "undefined") {
@@ -18,7 +7,38 @@ export const getBuildConfig = () => {
     );
   }
 
+  const buildMode = process.env.BUILD_MODE ?? "standalone";
+  const isApp = !!process.env.BUILD_APP;
+  const version = "v" + tauriConfig.package.version;
+
+  const commitInfo = (() => {
+    try {
+      const childProcess = require("child_process");
+      const commitDate: string = childProcess
+        .execSync('git log -1 --format="%at000" --date=unix')
+        .toString()
+        .trim();
+      const commitHash: string = childProcess
+        .execSync('git log --pretty=format:"%H" -n 1')
+        .toString()
+        .trim();
+
+      return { commitDate, commitHash };
+    } catch (e) {
+      console.error("[Build Config] No git or not from git repo.");
+      return {
+        commitDate: "unknown",
+        commitHash: "unknown",
+      };
+    }
+  })();
+
   return {
-    commitId: COMMIT_ID,
+    version,
+    ...commitInfo,
+    buildMode,
+    isApp,
   };
 };
+
+export type BuildConfig = ReturnType<typeof getBuildConfig>;
