@@ -1,10 +1,63 @@
-/** @type {import('next').NextConfig} */
+import webpack from "webpack";
 
+const mode = process.env.BUILD_MODE ?? "standalone";
+console.log("[Next] build mode", mode);
+
+const disableChunk = !!process.env.DISABLE_CHUNK || mode === "export";
+console.log("[Next] build with chunk: ", !disableChunk);
+
+/** @type {import('next').NextConfig} */
 const nextConfig = {
-  experimental: {
-    appDir: true,
+  webpack(config) {
+    config.module.rules.push({
+      test: /\.svg$/,
+      use: ["@svgr/webpack"],
+    });
+
+    if (disableChunk) {
+      config.plugins.push(
+        new webpack.optimize.LimitChunkCountPlugin({ maxChunks: 1 }),
+      );
+    }
+
+    config.resolve.fallback = {
+      child_process: false,
+    };
+
+    return config;
   },
-  async rewrites() {
+  output: mode,
+  images: {
+    unoptimized: mode === "export",
+  },
+};
+
+if (mode !== "export") {
+  nextConfig.headers = async () => {
+    return [
+      {
+        source: "/api/:path*",
+        headers: [
+          { key: "Access-Control-Allow-Credentials", value: "true" },
+          { key: "Access-Control-Allow-Origin", value: "*" },
+          {
+            key: "Access-Control-Allow-Methods",
+            value: "*",
+          },
+          {
+            key: "Access-Control-Allow-Headers",
+            value: "*",
+          },
+          {
+            key: "Access-Control-Max-Age",
+            value: "86400",
+          },
+        ],
+      },
+    ];
+  };
+
+  nextConfig.rewrites = async () => {
     const ret = [
       {
         source: "/api/proxy/:path*",
@@ -13,6 +66,10 @@ const nextConfig = {
       {
         source: "/google-fonts/:path*",
         destination: "https://fonts.googleapis.com/:path*",
+      },
+      {
+        source: "/sharegpt",
+        destination: "https://sharegpt.com/api/conversations",
       },
     ];
 
@@ -28,16 +85,7 @@ const nextConfig = {
     return {
       beforeFiles: ret,
     };
-  },
-  webpack(config) {
-    config.module.rules.push({
-      test: /\.svg$/,
-      use: ["@svgr/webpack"],
-    });
-
-    return config;
-  },
-  output: "standalone",
-};
+  };
+}
 
 export default nextConfig;
